requirements.txt
# requirements.txt

# Para a lógica principal e manipulação de dados (DataFrameService)
pandas

# Para a funcionalidade original do Google Drive (DriveService)
google-api-python-client
google-auth-oauthlib